Add `place` method to Robot class
Add a `place` method to the Robot class as required by Toy Robot
specification, that accepts a string argument that contains a new `x`,
`y` and `facing` coordinates; e.g. `"2,3,SOUTH"`; that is then split and
stored in their appropriate Robot variables, which updates the location
of the current Robot instance.

== References:
- [lua\-users wiki: Split Join]
  (http://lua-users.org/wiki/SplitJoin)

- [Lua 5\.2 Reference Manual]
  (http://www.lua.org/manual/5.2/manual.html#pdf-string.gmatch)

- [How do I get first table value in Lua \- Stack Overflow]
  (https://stackoverflow.com/questions/27850362/how-do-i-get-first-table-value-in-lua)

- [Lua string to int \- Stack Overflow]
  (https://stackoverflow.com/questions/10962085/lua-string-to-int)