Add `exec` method to Robot class
Add an execute method `exec` to the Robot class in order to call a Robot
instance method based off the given command as a string, and pass any
arguments to said command if they accept them; like `place`.

However, since there is no `switch` / `case` control flow structure in
Lua a table is used instead to contain an associative array of key value
pairs, the key being the raw Command name and the value being a callable
function that calls one of the methods on self. (Note: these methods need
to be in a code block / inline function, otherwise they get called
immediately on the spot).

== References:
- [What is the alternative for switch statement in Lua language? \- Stack
  Overflow]
  (https://stackoverflow.com/questions/37447704/what-is-the-alternative-for-switch-statement-in-lua-language)